requirements.txt
numpy==1.26.4
pandas==2.2.3
matplotlib==3.10.0
python-binance==1.0.27
python-dotenv==1.0.1
smartmoneyconcepts==0.0.24
seaborn==0.13.2
plotly==6.0.0
tqdm==4.67.1